Move sensor setting logic to MainViewModel
